+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # OneLineLogger Change Log
 
-### 1.0.1
+## 1.2.2
+
+* Logging level can be set by `setLevel(level)`, where parameter `level` is one of `logger.DEBUG`, `logger.INFO`, `logger.WARN`, `logger.ERROR`
+* `setGlobalDebugging()` is not deprecated. Use `setLevel(logger.DEBUG)` instead.
+* `logger.boo()` as an alias to `logger.highlight()`
+
+## 1.0.1
+
 * Added Install Instructions
 
 
-### 1.0.0
+## 1.0.0
+
 * Initial Release